requirements.txt
streamlit==1.32.0
plotly==5.20.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pandas==2.2.1
pymupdf==1.24.1
Pillow==10.3.0
streamlit-drawable-canvas==0.9.3
numpy==1.26.4
